backend_structure_adjustment: Grouped Services, created new trait for testing and fixed AI function services structure
