+++ amp-laravel/app/Http/Controllers/Admin/AdminController.php
@@ -2,23 +2,20 @@
 
 namespace App\Http\Controllers\Admin;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Services\Admin\ProviderService;
+use App\Services\Admin\ContactFormService;
+use App\Services\Admin\EditProfileService;
 use App\Http\Requests\Admin\EditProfileRequest;
 use App\Http\Requests\Admin\EditProviderRequest;
-use App\Services\Admin\editProfileService;
-use App\Services\Admin\getAllProvidersService;
-use App\Services\Admin\DeleteContactMessageService;
-use App\Services\Admin\getAllContactMessagesService;
-use App\Services\Admin\AdminEditProviderProfileService;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
-class AdminFunctionsController extends Controller
+class AdminController extends Controller
 {
     public function getProviders()
     {
         try {
-            $providers = getAllProvidersService::getAll();
+            $providers = ProviderService::getAll();
 
             return $this->messageResponse(true, "Providers retrieved successfully", 200, $providers);
         } catch (\Exception $e) {
@@ -26,10 +23,21 @@ class AdminFunctionsController extends Controller
         }
     }
 
+    public function editProvider(EditProviderRequest $request, $id)
+    {
+        try {
+            ProviderService::editProfile($id, $request->validated());
+
+            return $this->messageResponse(true, "Profile updated successfully", 200);
+        } catch (\Exception $e) {
+            return $this->errorMessageResponse(false, $e->getMessage(), "Failed to update profile", 500);
+        }
+    }
+
     public function getContactMessages()
     {
         try {
-            $messages = getAllContactMessagesService::getAll();
+            $messages = ContactFormService::getAll();
 
             return $this->messageResponse(true, "Messages retrieved successfully", 200, $messages);
         } catch (\Exception $e) {
@@ -37,39 +45,27 @@ class AdminFunctionsController extends Controller
         }
     }
 
-    public function editProvider(EditProviderRequest $request, $id)
-    {
-        try {
-            AdminEditProviderProfileService::editProfile($id, $request->validated());
-
-            return $this->messageResponse(true, "Profile updated successfully", 200);
-        } catch (\Exception $e) {
-            return $this->errorMessageResponse(false, $e->getMessage(), "Failed to update profile", 500);
-        }
-    }
-
-
-    public function editProfile(EditProfileRequest $request)
-    {
-        try {
-            editProfileService::editProfile($request->validated());
-
-            return $this->messageResponse(true, "Profile updated successfully", 200);
-        } catch (\Exception $e) {
-            return $this->errorMessageResponse(false, $e->getMessage(), "Failed to update profile", 500);
-        }
-    }
-
     public function deleteMessage($id)
     {
         try {
-            DeleteContactMessageService::delete($id);
+            ContactFormService::delete($id);
 
             return $this->messageResponse(true, "Message deleted successfully", 200);
         } catch (ModelNotFoundException $e) {
             return $this->errorMessageResponse(false, "ID not Found", $e->getMessage(), 404);
         } catch (\Exception $e) {
             return $this->errorMessageResponse(false, $e->getMessage(), "Failed to delete message", 500);
+        }
+    }
+
+    public function editProfile(EditProfileRequest $request)
+    {
+        try {
+            EditProfileService::editProfile($request->validated());
+
+            return $this->messageResponse(true, "Profile updated successfully", 200);
+        } catch (\Exception $e) {
+            return $this->errorMessageResponse(false, $e->getMessage(), "Failed to update profile", 500);
         }
     }
 }
