--- amp-laravel/app/Services/Admin/getAllContactMessagesService.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Services\Admin;
-
-use App\Models\ContactForm;
-
-class getAllContactMessagesService
-{
-    public static function getAll()
-    {
-        return ContactForm::all();
-    }
-}
