+++ amp-laravel/app/Traits/ResponseTrait.php
@@ -2,9 +2,7 @@
 
 namespace App\Traits;
 
-use App\Models\User;
 use Illuminate\Contracts\Validation\Validator;
-use PHPOpenSourceSaver\JWTAuth\Facades\JWTAuth;
 use Illuminate\Http\Exceptions\HttpResponseException;
 
 trait ResponseTrait
@@ -43,29 +41,5 @@ trait ResponseTrait
             'errors' => $validator->errors(),
             'message' => 'Invalid Credentials'
         ], 422));
-    }
-
-    public function actingAsClient()
-    {
-        $client = User::factory()->create(['user_type' => 'Client']);
-        $token = JWTAuth::fromUser($client);
-        $request = $this->withHeader('Authorization', "Bearer $token");
-
-        return [$request, $client];
-    }
-
-
-    public function actingAsProvider()
-    {
-        $provider = User::factory()->create(['user_type' => 'provider']);
-        $token = JWTAuth::fromUser($provider);
-        return $this->withHeader('Authorization', "Bearer $token");
-    }
-
-    public function actingAsAdmin()
-    {
-        $admin = User::factory()->create(['user_type' => 'admin']);
-        $token = JWTAuth::fromUser($admin);
-        return $this->withHeader('Authorization', "Bearer $token");
     }
 }
