+++ amp-laravel/app/Utils/ClientPrompt.php
@@ -1,18 +1,13 @@
 <?php
 
-namespace App\Services\Client;
+namespace App\Utils;
 
+use Exception;
 use App\Models\Metric;
 
-use Prism\Prism\Prism;
-use Prism\Prism\Enums\Provider;
-
-use Prism\Prism\Schema\ObjectSchema;
-use Prism\Prism\Schema\StringSchema;
-
-class GeneratingReportService
+class ClientPrompt
 {
-    public static function generateReport($slaveId)
+    public static function clientBuildPromptFromMetrics($slaveId): string
     {
         $metrics = Metric::where('slave_id', $slaveId)
             ->orderBy('date_month', 'desc')
@@ -44,32 +39,6 @@ class GeneratingReportService
                     $metricsText
                     EOT;
 
-
-        $schema = new ObjectSchema(
-            name: 'metrics_report',
-            description: 'A structured energy metrics report for a slave device',
-            properties: [
-                new StringSchema('summary', 'Overall performance summary'),
-                new StringSchema('voltageInsights', 'Insights on voltage trends'),
-                new StringSchema('powerInsights', 'Insights on power consumption'),
-                new StringSchema('energyInsights', 'Insights on energy usage'),
-                new StringSchema('recommendations', 'Recommended actions based on data')
-            ],
-            requiredFields: ['summary', 'voltageInsights', 'powerInsights', 'energyInsights']
-        );
-
-        $response = Prism::structured()
-        // ->using(Provider::OpenAI, 'o4-mini')
-            ->using(Provider::OpenAI, 'gpt-4.1')
-            ->withSchema($schema)
-            ->withPrompt($prompt)
-            ->asStructured();
-
-        $structured = $response->structured;
-
-        $structured['recommendations'] = str_replace('\n', "\n", $structured['recommendations']);
-        $structured['summary'] = preg_replace('/device\s+\d+/i', 'the device', $structured['summary']);
-
-        return $structured;
+        return $prompt;
     }
 }
