+++ amp-laravel/database/factories/ContactFormFactory.php
@@ -18,7 +18,6 @@ class ContactFormFactory extends Factory
     public function definition(): array
     {
         return [
-            'user_id' => null,
             'name' => $this->faker->name,
             'email' => $this->faker->unique()->safeEmail,
             'phone_number' => $this->faker->phoneNumber,
