+++ amp-laravel/database/seeders/ContactFormSeeder.php
@@ -2,7 +2,6 @@
 
 namespace Database\Seeders;
 
-use App\Models\User;
 use App\Models\ContactForm;
 use Illuminate\Database\Seeder;
 
@@ -10,12 +9,6 @@ class ContactFormSeeder extends Seeder
 {
     public function run(): void
     {
-        $users = User::inRandomOrder()->limit(10)->get();
-
-        foreach ($users as $user) {
-            ContactForm::factory()->create([
-                'user_id' => $user->id,
-            ]);
-        }
+        ContactForm::factory()->count(10)->create();
     }
 }
