+++ amp-laravel/tests/Feature/MetricTest.php
@@ -5,12 +5,12 @@ namespace Tests\Feature;
 use Tests\TestCase;
 use App\Models\Slave;
 use App\Models\Master;
-use App\Traits\ResponseTrait;
+use App\Traits\TestTrait;
 use Illuminate\Foundation\Testing\WithFaker;
 
 class MetricTest extends TestCase
 {
-    use WithFaker, ResponseTrait;
+    use WithFaker, TestTrait;
 
     public function testSubmitSlaveMetricsSuccessfully(): void
     {
